% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{adjust.covariates}
\alias{adjust.covariates}
\title{Adjust exon length}
\usage{
adjust.covariates(x, offset, group)
}
\arguments{
\item{x}{matrix with \eqn{n} rows (samples) and \eqn{p} columns (exons)}

\item{offset}{exon start positions\strong{:} vector of length \eqn{p}
exon end positions\strong{:} vector of length \eqn{p}}

\item{group}{gene (not exon) names\strong{:} vector of length \eqn{p}}
}
\description{
This function adjusts exon expression data for different exon lengths.
}
\details{
No information on chromosomes required.
}
\examples{
NA

}
