% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{adjust.exon.length}
\alias{adjust.exon.length}
\title{Adjust exon length}
\usage{
adjust.exon.length(y, gene, start, end)
}
\arguments{
\item{y}{matrix with \eqn{n} rows (samples) and \eqn{p} columns (exons)}

\item{gene}{gene (not exon) names\strong{:} vector of length \eqn{p}}

\item{start}{exon start positions\strong{:} vector of length \eqn{p}}

\item{end}{exon end positions\strong{:} vector of length \eqn{p}}
}
\description{
This function adjusts exon expression data for different exon lengths.
}
\details{
No information on chromosomes required.
}
\examples{
NA

}
