% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{adjust.library.sizes}
\alias{adjust.library.sizes}
\title{Adjust library sizes}
\usage{
adjust.library.sizes(Y)
}
\arguments{
\item{Y}{matrix with \eqn{n} rows (samples) and \eqn{p} columns (exons)}
}
\description{
This function adjusts exon expression data for different library sizes.
}
\examples{
NA

}
