% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{adjust.samples}
\alias{adjust.samples}
\title{Adjust library sizes}
\usage{
adjust.samples(x)
}
\arguments{
\item{x}{matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}
}
\description{
This function adjusts RNA-seq expression data for different library sizes.
}
\examples{
n <- 5; p <- 10
x <- matrix(rnbinom(n=n*p,mu=5,size=1/0.5),nrow=n,ncol=p)
x[1,] <- 10*x[1,]
adjust.samples(x)

}
