% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilasso.R
\name{bilasso}
\alias{bilasso}
\title{bilasso}
\usage{
bilasso(y, cutoff, X, alpha = 1, nfolds = 10)
}
\arguments{
\item{y}{continuous response\strong{:}
vector of length \eqn{n}}

\item{cutoff}{value between \code{min(y)} and \code{max(y)}}

\item{X}{covariates\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{alpha}{elastic net parameter\strong{:}
numeric between \eqn{0} and \eqn{1};
\eqn{alpha=1} for lasso,
\eqn{alpha=0} for ridge}

\item{nfolds}{number of folds}

\item{z}{binary response\strong{:}
vector of length \eqn{n}}
}
\description{
Implements penalised regression with response duality.
\code{pi=0} represents binomial regression,
\code{pi=1} represents linear regression
}
\examples{
NA

}
