% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilasso.R
\name{bilasso_compare}
\alias{bilasso_compare}
\title{Comparison}
\usage{
bilasso_compare(y, cutoff, X, nfolds = 5)
}
\arguments{
\item{y}{continuous response\strong{:}
vector of length \eqn{n}}

\item{cutoff}{cutoff point for dichotomising response into classes\strong{:}
value between \code{min(y)} and \code{max(y)}}

\item{X}{covariates\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{nfolds}{number of folds}
}
\description{
Compares models for a continuous response with a cutoff value
}
\examples{
NA

}
