% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.mixnet}
\alias{coef.mixnet}
\title{Extract Coefficients}
\usage{
\method{coef}{mixnet}(object, ...)
}
\arguments{
\item{object}{\link[mixnet]{mixnet} object}

\item{...}{further arguments (not applicable)}
}
\description{
Extracts pooled coefficients.
(The meta learners weights the coefficients from the base learners.)
}
\examples{
n <- 30; q <- 2; p <- 20
Y <- matrix(rnorm(n*q),nrow=n,ncol=q)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
object <- mixnet(Y=Y,X=X)
coef <- coef(object)

}
