% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_compare}
\alias{colasso_compare}
\title{External cross-validation}
\usage{
colasso_compare(y, X, plot = TRUE)
}
\arguments{
\item{plot}{logical}

\item{n}{sample size}

\item{p}{number of covariates}

\item{cor}{correlation structure}
}
\description{
This function ...
}
\examples{
NA

}
