% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_covariate_weights}
\alias{colasso_covariate_weights}
\title{covariate weights}
\usage{
colasso_covariate_weights(x, max = 0.05/length(x), min = 0.05,
  version = 1)
}
\arguments{
\item{x}{\eqn{p}-values\strong{:}
vector with entries between \eqn{0} and \eqn{1}}

\item{max}{maximum \eqn{p}-value to receive weight one}

\item{min}{minimum \eqn{p}-value to receive weight zero}

\item{version}{(temporary argument)}
}
\description{
This function ...
}
\examples{
# alternative weights: weight <- abs(cor(y,X))

}
