% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_marginal_significance}
\alias{colasso_marginal_significance}
\title{marginal significance}
\usage{
colasso_marginal_significance(y, X)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{X}{covariates\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}
}
\description{
This function ...
}
\examples{
NA

}
