% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_moderate}
\alias{colasso_moderate}
\title{moderated response}
\usage{
colasso_moderate(y, X, pi, plot = FALSE)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{X}{covariates\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{pi}{vector with entries between \eqn{0} and \eqn{1} (rename argument)}

\item{plot}{logical}
}
\description{
This function ...
}
\examples{
NA

}
