% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_simulate}
\alias{colasso_simulate}
\title{simulate data}
\usage{
colasso_simulate(n = 100, p = 500, cor = "constant", plot = TRUE)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{cor}{correlation structure}

\item{plot}{logical}
}
\description{
This function ...
}
\examples{
# CONTINUE HERE

}
