% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_weighted_correlation}
\alias{colasso_weighted_correlation}
\title{weighted correlation}
\usage{
colasso_weighted_correlation(X, w = NULL)
}
\arguments{
\item{X}{covariates\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{w}{covariate weights\strong{:}
vector of length \eqn{p}, with entries between \eqn{0} and \eqn{1}}
}
\description{
This function ...
}
\examples{
NA

}
