% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.check}
\alias{.check}
\title{Arguments}
\usage{
.check(x, type, miss = FALSE, min = NULL, max = NULL,
  values = NULL, inf = FALSE, null = FALSE)
}
\arguments{
\item{x}{argument}

\item{type}{character \code{"string"}, \code{"scalar"}, \code{"vector"}, \code{"matrix"}}

\item{miss}{accept missing values\strong{:}
logical}

\item{min}{lower limit\strong{:}
numeric}

\item{max}{upper limit\strong{:}
numeric}

\item{values}{only accept specific values\strong{:}
vector}

\item{inf}{accept infinite (\code{Inf} or \code{-Inf}) values\strong{:}
logical}

\item{null}{accept \code{NULL}\strong{:}
logical}
}
\description{
Verifies whether an argument matches formal requirements.
}
\examples{
cornet:::.check(0.5,type="scalar",min=0,max=1)

}
