% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.compare}
\alias{.compare}
\title{Comparison}
\usage{
.compare(y, cutoff, X, alpha = 1, nfolds = 5, foldid = NULL,
  type.measure = "deviance", trial = FALSE)
}
\arguments{
\item{y}{continuous response\strong{:}
vector of length \eqn{n}}

\item{cutoff}{cutoff point for dichotomising response into classes\strong{:}
value between \code{min(y)} and \code{max(y)}}

\item{X}{covariates\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{alpha}{elastic net mixing parameter\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{nfolds}{number of folds}

\item{foldid}{fold identifiers\strong{:}
vector with entries between \eqn{1} and \code{nfolds};
or \code{NULL} (balance)}

\item{type.measure}{loss function for binary classification
(linear regression uses the deviance)}

\item{trial}{logical}
}
\description{
Compares models for a continuous response with a cutoff value
}
\examples{
NA

}
