% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.equal}
\alias{.equal}
\title{Equality}
\usage{
.equal(..., na.rm = FALSE)
}
\arguments{
\item{...}{scalars, vectors, or matrices of equal dimensions}

\item{na.rm}{remove missing values\strong{:}
logical}
}
\description{
Verifies whether two or more arguments are identical.
}
\examples{
cornet:::.equal(1,1,1)

}
