% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.simulate}
\alias{.simulate}
\title{Data simulation}
\usage{
.simulate(n, p, prob = 0.2, fac = 1)
}
\arguments{
\item{n}{sample size\strong{:}
positive integer}

\item{p}{covariate space\strong{:}
positive integer}

\item{prob}{(approximate) proportion of causal covariates\strong{:}
numeric between \eqn{0} and \eqn{1}}

\item{fac}{noise factor\strong{:}
positive real number}
}
\value{
Returns invisible list with elements \code{y} and \code{X}.
}
\description{
Simulates data for unit tests
}
\examples{
data <- cornet:::.simulate(n=10,p=20,prob=0.2,fac=2)
names(data)

}
