% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop.trivial}
\alias{drop.trivial}
\title{Drop trivial tests}
\usage{
drop.trivial(map)
}
\arguments{
\item{map}{list with names "genes", "exons", and "snps"
(output from \code{\link{map.genes}}, \code{\link{map.exons}},
and \code{\link{map.snps}})}
}
\description{
This function trops trivial tests.
}
\details{
This functions drops tests for genes without SNPs or with a single exon.
}
\examples{
NA

}
