% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.exons.bbmri}
\alias{get.exons.bbmri}
\title{Get exon data (BBMRI)}
\usage{
get.exons.bbmri(path = getwd())
}
\arguments{
\item{path}{data directory}
}
\description{
This function transforms exon data (virtual machine).
}
\examples{
path <- "/virdir/Scratch/arauschenberger/trial"

}
