% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.snps.geuvadis}
\alias{get.snps.geuvadis}
\title{Get SNP data (Geuvadis)}
\usage{
get.snps.geuvadis(chr, data = NULL, path = getwd())
}
\arguments{
\item{chr}{chromosome: integer \eqn{1-22}}

\item{data}{local directory for VCF (variant call format)
and SDRF (sample and data relationship format) files}

\item{path}{local directory for output}
}
\description{
This function transforms SNP data (local machine).
}
\examples{
path <- "C:/Users/a.rauschenbe/Desktop/spliceQTL/data"

}
