% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grid}
\alias{grid}
\title{Plot grid}
\usage{
grid(x, y, n = 10, ...)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{n}{grid resolution}

\item{...}{graphical parameters}
}
\description{
This function combines a scatter plot with a density plot.
}
\examples{
x <- stats::rbeta(n=100,shape1=0.3,shape2=0.5)
y <- stats::rbeta(n=100,shape1=0.3,shape2=0.5)
grid(x,y)

}
