% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{joinet}
\alias{joinet}
\title{Multivariate Elastic Net Regression}
\usage{
joinet(
  Y,
  X,
  family = "gaussian",
  nfolds = 10,
  foldid = NULL,
  type.measure = "deviance",
  alpha.base = 1,
  alpha.meta = 1,
  constraint = TRUE,
  weight = NULL,
  ...
)
}
\arguments{
\item{Y}{outputs\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{q} columns (variables)}

\item{X}{inputs\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{family}{distribution\strong{:}
vector of length \eqn{1} or \eqn{q} with entries
\code{"gaussian"}, \code{"binomial"} or \code{"poisson"}}

\item{nfolds}{number of folds}

\item{foldid}{fold identifiers\strong{:}
vector of length \eqn{n} with entries between \eqn{1} and \code{nfolds};
or \code{NULL} (balance)}

\item{type.measure}{loss function\strong{:}
vector of length \eqn{1} or \eqn{q} with entries
\code{"deviance"}, \code{"class"}, \code{"mse"} or \code{"mae"}
(see \code{\link[glmnet]{cv.glmnet}})}

\item{alpha.base}{elastic net mixing parameter for base learners\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{alpha.meta}{elastic net mixing parameter for meta learners\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{constraint}{non-negativity constraints\strong{:}
logical (see details)}

\item{weight}{inclusion/exclusion of variables\strong{:}
logical matrix with \eqn{q} rows and \eqn{p} columns,
or \code{NULL} (see details)}

\item{...}{further arguments passed to \code{\link[glmnet]{glmnet}}}
}
\value{
This function returns an object of class \code{joinet}.
Available methods include
\code{\link[=predict.joinet]{predict}},
\code{\link[=coef.joinet]{coef}},
and \code{\link[=weights.joinet]{weights}}.
The slots \code{base} and \code{meta} each contain
\eqn{q} \code{\link[glmnet]{cv.glmnet}}-like objects.
}
\description{
Implements multivariate elastic net regression.
}
\details{
\strong{non-negativity constraints:}
If it is reasonable to assume that the outcomes
are \emph{positively} correlated
(potentially after changing the sign of some outcomes)
we recommend to set \code{constraint=TRUE}.
Then non-negativity constraints are imposed on the meta learner.

\strong{inclusion/exclusion of variables:}
The entry in the \eqn{j}th column and the \eqn{k}th row
indicates whether the \eqn{j}th feature may be used for 
modelling the \eqn{k}th outcome
(where \eqn{0} means \code{FALSE} and
\eqn{1} means \code{TRUE}).

\strong{elastic net:}
\code{alpha.base} controls input-output effects,
\code{alpha.meta} controls output-output effects;
lasso renders sparse models (\code{alpha}\eqn{=1}),
ridge renders dense models (\code{alpha}\eqn{=0})
}
\examples{
\dontshow{
if(!grepl('SunOS',Sys.info()['sysname'])){
n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
object <- joinet(Y=Y,X=X)}}
\dontrun{
n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
object <- joinet(Y=Y,X=X)}

\dontrun{
browseVignettes("joinet") # further examples}

}
\references{
Armin Rauschenberger, Enrico Glaab (2021)
"Predicting correlated outcomes from molecular data"
\emph{Manuscript in preparation}.
}
\seealso{
\code{\link{cv.joinet}}, vignette
}
