% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map.exons}
\alias{map.exons}
\title{Search for exons}
\usage{
map.exons(gene, exon)
}
\arguments{
\item{gene}{gene names\strong{:} vector with one entry per gene,
including the gene names}

\item{exon}{exon names\strong{:} vector with one entry per exon,
including the corresponding \emph{gene} names
(separated by comma if multiple gene names)}
}
\description{
This function attributes exons to genes.
}
\details{
The exon names should contain the gene names. For each gene, this function
returns the indices of the exons.
}
\examples{
gene <- c("A","B","C")
exon <- c("A","A,B","B","B,C","C")
map.exons(gene,exon)

}
