% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map.genes}
\alias{map.genes}
\title{Search for genes}
\usage{
map.genes(chr, path = getwd(), release = "GRCh37", build = 71)
}
\arguments{
\item{chr}{chromosome\strong{:} integer 1-22}

\item{path}{path to gene transfer format files (.gtf)}

\item{release}{character "NCBI36", "GRCh37", or "GRCh38"}

\item{build}{integer 49-91}
}
\description{
This function retrieves all genes on a chromosome.
}
\examples{
NA

}
