% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{map.snps}
\alias{map.snps}
\title{Search for SNPs}
\usage{
map.snps(gene.chr, gene.start, gene.end, snp.chr, snp.pos)
}
\arguments{
\item{gene.chr}{chromosome\strong{:}
numeric vector with one entry per gene}

\item{gene.start}{start position\strong{:}
numeric vector with one entry per gene}

\item{gene.end}{end position\strong{:}
numeric vector with one entry per gene}

\item{snp.chr}{integer 1-22}

\item{snp.pos}{chromosomal position of SNPs\strong{:}
numeric vector with one entry per SNP}
}
\description{
This function
}
\details{
This functions ...
}
\examples{
NA

}
