% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{match.samples}
\alias{match.samples}
\title{Prepare data matrices}
\usage{
match.samples(..., message = TRUE)
}
\arguments{
\item{...}{matrices with samples in the rows and variables in the columns
(separated by commas)}

\item{message}{display messages\strong{:} logical}
}
\description{
This function removes duplicate samples from each matrix,
only retains samples appearing in all matrices,
and brings them into the same order.
}
\examples{
NA

}
