% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mixnet}
\alias{mixnet}
\title{Multivariate Elastic Net Regression}
\usage{
mixnet(Y, X, family = "gaussian", nfolds = 10, foldid = NULL,
  type.measure = "deviance", alpha.base = 0, alpha.meta = 0, ...)
}
\arguments{
\item{Y}{outputs\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{q} columns (variables),
with positive correlation (see details)}

\item{X}{inputs\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{family}{distribution\strong{:}
vector of length \eqn{1} or \eqn{q} with entries
\code{"gaussian"}, \code{"binomial"} or \code{"poisson"}}

\item{nfolds}{number of folds}

\item{foldid}{fold identifiers\strong{:}
vector of length \eqn{n} with entries between \eqn{1} and \code{nfolds};
or \code{NULL} (balance)}

\item{type.measure}{loss function\strong{:}
vector of length \eqn{1} or \eqn{q} with entries
\code{"deviance"}, \code{"class"}, \code{"mse"} or \code{"mae"}
(see \code{\link[glmnet]{cv.glmnet}})}

\item{alpha.base}{elastic net mixing parameter for base learners\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{alpha.meta}{elastic net mixing parameter for meta learner\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{...}{further arguments passed to \code{\link[glmnet]{glmnet}}}
}
\description{
Implements multivariate elastic net regression.
}
\details{
The \eqn{q} outcomes should be positively correlated.
Avoid negative correlations by changing the sign of the variable.

elastic net mixing parameters:
\code{alpha.base} controls input-output effects,
\code{alpha.meta} controls output-output effects;
ridge (\eqn{0}) renders dense models,
lasso (\eqn{1}) renders sparse models
}
\examples{
n <- 30; q <- 2; p <- 20
Y <- matrix(rnorm(n*q),nrow=n,ncol=q)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
object <- mixnet(Y=Y,X=X)

}
\references{
A Rauschenberger, E Glaab (2019)
"Multivariate elastic net regression through stacked generalisation"
\emph{Manuscript in preparation.}
}
