% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colasso.R
\name{moderate}
\alias{moderate}
\title{moderated response (trial)}
\usage{
moderate(y, Y, k = 2)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{Y}{response\strong{:}
matrix with \eqn{n} rows and \eqn{p} columns,
or vector of length \eqn{n} (see details)}
}
\description{
This function ...
}
\examples{
NA

}
