% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.cornet}
\alias{predict.cornet}
\title{to do}
\usage{
\method{predict}{cornet}(object, newx, type = "probability", ...)
}
\arguments{
\item{object}{cornet object}

\item{newx}{covariates\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{type}{\code{"probability"}, \code{"odds"}, \code{"log-odds"}}

\item{...}{to do}
}
\description{
to do
}
\examples{
NA

}
