% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.mixnet}
\alias{predict.mixnet}
\title{Makes Predictions}
\usage{
\method{predict}{mixnet}(object, newx, type = "response", ...)
}
\arguments{
\item{object}{\link[mixnet]{mixnet} object}

\item{newx}{covariates\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{type}{character "link" or "response"}

\item{...}{further arguments (not applicable)}
}
\description{
Predicts outcome from features with stacked model.
}
\examples{
n <- 30; q <- 2; p <- 20
Y <- matrix(rnorm(n*q),nrow=n,ncol=q)
Y <- matrix(rbinom(n=n*q,size=1,prob=0.5),nrow=n,ncol=q)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
object <- mixnet(Y=Y,X=X,family="binomial")
y_hat <- predict(object,newx=X)

}
