% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{prepare.data.matrices}
\alias{prepare.data.matrices}
\title{Prepare data matrices}
\usage{
prepare.data.matrices(Y, X)
}
\arguments{
\item{Y}{matrix with \eqn{n_y} rows (samples) and \eqn{p_y} columns (exons)}

\item{X}{matrix with \eqn{n_x} rows (samples) and \eqn{p_x} columns (SNPs)}
}
\description{
This function removes duplicate samples, retains overlapping samples, and
orders samples.
}
\examples{
NA

}
