% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{spliceQTL-package}
\alias{spliceQTL-package}
\alias{spliceQTL}
\title{Alternative Splicing}
\description{
This R package includes various functions
for applying the global test of alternative splicing.
Some functions only work on the virtual machine (see below).
}
\seealso{
Prepare BBMRI and Geuvadis data:
\itemize{
\item \code{\link{get.snps.geuvadis}} (not VM)
\item \code{\link{get.snps.bbmri}} (only VM)
\item \code{\link{get.exons.geuvadis}} (only VM)
\item \code{\link{get.exons.bbmri}} (only VM)
}

Process samples and covariates:
\itemize{
\item \code{\link{match.samples}}
\item \code{\link{adjust.samples}}
\item \code{\link{adjust.covariates}}
}

Search for exons and SNPs:
\itemize{
\item \code{\link{map.genes}}
\item \code{\link{map.exons}}
\item \code{\link{map.snps}}
\item \code{\link{drop.trivial}}
}

Test for alternative splicing:
\itemize{
\item \code{\link{test.single}}
\item \code{\link{test.multiple}}
}
}
\keyword{documentation}
