% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{spliceQTL}
\alias{spliceQTL}
\title{Conduct tests}
\usage{
spliceQTL(Y, X, map, i, limit, steps)
}
\arguments{
\item{Y}{exon expression\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (exons)}

\item{X}{SNP genotype\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{q} columns (SNPs)}

\item{map}{list with names "genes", "exons", and "snps"
(output from \code{map.genes}, \code{map.exons}, and \code{map.snps})}

\item{i}{gene index\strong{:}
integer between \eqn{1} and \code{nrow(map$genes)}}

\item{limit}{cutoff for rounding \code{p}-values}

\item{steps}{size of permutation chunks\strong{:}
integer vector}
}
\description{
This function
}
\details{
The maximum number of permutations equals \code{sum(steps)}. Permutations is
interrupted if at least \code{limit} test statistics for the permuted data
are larger than the test statistic for the observed data.
}
\examples{
NA

}
