% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{visualise}
\alias{visualise}
\title{Plot SNP-exon correlations}
\usage{
visualise(Y, X, map, i)
}
\arguments{
\item{Y}{exon expression\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (exons)}

\item{X}{SNP genotype\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{q} columns (SNPs)}

\item{map}{list with names "genes", "exons", and "snps"
(output from \code{map.genes}, \code{map.exons}, and \code{map.snps})}

\item{i}{gene index\strong{:}
integer between \eqn{1} and \code{nrow(map$genes)}}
}
\description{
This function ...
}
\examples{
# see vignette

}
