% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weights.joinet}
\alias{weights.joinet}
\title{Extract Weights}
\usage{
\method{weights}{joinet}(object, ...)
}
\arguments{
\item{object}{\link[joinet]{joinet} object}

\item{...}{further arguments (not applicable)}
}
\description{
Extracts coefficients from the meta learner,
i.e. the weights for the base learners.
}
\examples{
n <- 30; q <- 2; p <- 20
Y <- matrix(rnorm(n*q),nrow=n,ncol=q)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
object <- joinet(Y=Y,X=X)
weights(object)

}
