% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{adjust.samples}
\alias{adjust.samples}
\title{Adjust library sizes}
\usage{
adjust.samples(x)
}
\arguments{
\item{x}{matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}
}
\description{
This function adjusts RNA-seq expression data for different library sizes.
}
\examples{
NA

}
