% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilasso.R
\name{bilasso}
\alias{bilasso}
\title{Logistic regression with a continuous response}
\usage{
bilasso(y, cutoff, X, alpha = 1, nfolds = 10, foldid = NULL,
  type.measure = "deviance", nsigma = 100, sigma = NULL,
  logistic = TRUE)
}
\arguments{
\item{y}{continuous response\strong{:}
vector of length \eqn{n}}

\item{cutoff}{cutoff point for dichotomising response into classes\strong{:}
value between \code{min(y)} and \code{max(y)}}

\item{X}{covariates\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{alpha}{elastic net parameter\strong{:}
numeric between \eqn{0} (ridge)
and \eqn{1} (lasso)}

\item{nfolds}{number of folds}

\item{foldid}{fold identifiers\strong{:}
vector with entries between \eqn{1} and \code{nfolds};
or \code{NULL} (balance)}

\item{type.measure}{loss function for binary classification
(linear regression uses the deviance)}

\item{nsigma}{number of \code{sigma} values}

\item{sigma}{sigma sequence\strong{:}
vector of increasing positive values;
or \code{NULL} (default sequence)}

\item{logistic}{fit logistic regression for comparison\strong{:}
logical
(currently all methods require \code{TRUE})}
}
\description{
Implements logistic regression with a continuous response.
}
\details{
INCLUDE note on deviance (not comparable between lin and log models)
}
\examples{
n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
bilasso(y=y,cutoff=0,X=X)

}
