% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilasso.R
\name{bilasso}
\alias{bilasso}
\title{Logistic regression with a continuous response}
\usage{
bilasso(y, cutoff, X, alpha = 1, nfolds = 10,
  type.measure = "deviance")
}
\arguments{
\item{y}{continuous response\strong{:}
vector of length \eqn{n}}

\item{cutoff}{value between \code{min(y)} and \code{max(y)}}

\item{X}{covariates\strong{:}
matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{alpha}{elastic net parameter\strong{:}
numeric between \eqn{1} and \eqn{0};
compromise between lasso (\eqn{alpha=1})
and ridge (\eqn{alpha=0}) regression}

\item{nfolds}{number of folds}

\item{type.measure}{loss function for logistic regression
(the deviance is used for linear regression)}
}
\description{
Implements penalised logistic regression
with both a binary and a continuous response.
}
\details{
Finds a compromise between binomial (\eqn{pi=0})
and linear (\eqn{pi=1}) regression.
}
\examples{
NA

}
