% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso}
\alias{colasso}
\alias{colasso-package}
\title{colasso}
\usage{
colasso(y, X, nfold = 5, alpha = 1)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{X}{covariates\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{nfold}{number of folds}

\item{alpha}{elastic net parameter}
}
\description{
This function ...
}
\examples{
n <- 100; p <- 20
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
#y <- rbinom(n=n,size=1,prob=0.2)
y <- rnorm(n=n)
#y[1] <- 0.5
#a <- glmnet::glmnet(y=y,x=x,family="binomial")
#b <- stats::glm(y~x,family="binomial")
}
