% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colasso.R
\name{colasso}
\alias{colasso}
\alias{colasso-package}
\title{colasso}
\usage{
colasso(y, Y, X, alpha = 1, nfolds = 10, family = "gaussian",
  type.measure = "deviance")
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{Y}{response\strong{:}
matrix with \eqn{n} rows and \eqn{p} columns,
or vector of length \eqn{n} (see details)}

\item{X}{covariates\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{alpha}{elastic net parameter\strong{:}
numeric between \eqn{0} and \eqn{1};
\eqn{alpha=1} for lasso,
\eqn{alpha=0} for ridge}

\item{nfolds}{number of folds}

\item{family}{see glmnet}

\item{type.measure}{see glmnet}
}
\description{
Implements penalised regression with response moderation.
}
\examples{
n <- 100; p <- 20; q <- 10
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- matrix(rnorm(n*q),nrow=n,ncol=q)
#y <- rbinom(n=n,size=1,prob=0.2)
y <- rnorm(n=n)
test <- colasso(y=y,Y=Y,X=X)

}
