% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colasso.R
\name{colasso_compare}
\alias{colasso_compare}
\title{External cross-validation}
\usage{
colasso_compare(y, Y, X, plot = TRUE, nfolds.ext = 5,
  nfolds.int = 10, family = "gaussian", type.measure = "deviance")
}
\arguments{
\item{y}{response vector}

\item{Y}{response moderation matrix}

\item{X}{covariates}

\item{plot}{logical}

\item{nfolds.ext}{external folds}

\item{nfolds.int}{internal folds}

\item{family}{see glmnet}

\item{type.measure}{see glmnet}
}
\description{
This function ...
}
\examples{
NA

}
