% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_compare}
\alias{colasso_compare}
\title{External cross-validation}
\usage{
colasso_compare(y, X, plot = TRUE)
}
\arguments{
\item{y}{response}

\item{X}{covariates}

\item{plot}{logical}
}
\description{
This function ...
}
\examples{
NA

}
