% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colasso_compare}
\alias{colasso_compare}
\title{External cross-validation}
\usage{
colasso_compare(y, Y, X, plot = TRUE, nfolds.int = 10,
  family = "gaussian", type.measure = "deviance")
}
\arguments{
\item{y}{response vector}

\item{Y}{response moderation matrix}

\item{X}{covariates}

\item{plot}{logical}

\item{nfolds.int}{internal folds}
}
\description{
This function ...
}
\examples{
NA

}
