% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colasso.R
\name{colasso_moderate}
\alias{colasso_moderate}
\title{moderated response}
\usage{
colasso_moderate(Y, ...)
}
\arguments{
\item{Y}{response\strong{:}
matrix with \eqn{n} rows and \eqn{p} columns,
or vector of length \eqn{n} (see details)}

\item{...}{further arguments (currently not implemented)
vector with entries between \eqn{0} and \eqn{1} (rename argument)}
}
\description{
This function ...
}
\examples{
NA
}
