% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cornet}
\alias{cornet}
\title{Logistic regression with a continuous response}
\usage{
cornet(y, cutoff, X, alpha = 1, npi = 101, pi = NULL, nsigma = 99,
  sigma = NULL, nfolds = 10, foldid = NULL,
  type.measure = "deviance", ...)
}
\arguments{
\item{y}{continuous response\strong{:}
vector of length \eqn{n}}

\item{cutoff}{cutoff point for dichotomising response into classes\strong{:}
value between \code{min(y)} and \code{max(y)}}

\item{X}{covariates\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{alpha}{elastic net mixing parameter\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{npi}{number of \code{pi} values}

\item{pi}{pi sequence\strong{:}
vector of values in the unit interval;
or \code{NULL} (default sequence)}

\item{nsigma}{number of \code{sigma} values}

\item{sigma}{sigma sequence\strong{:}
vector of increasing positive values;
or \code{NULL} (default sequence)}

\item{nfolds}{number of folds}

\item{foldid}{fold identifiers\strong{:}
vector with entries between \eqn{1} and \code{nfolds};
or \code{NULL} (balance)}

\item{type.measure}{loss function for binary classification
(linear regression uses the deviance)}

\item{...}{further arguments passed to \code{\link[glmnet]{glmnet}}}
}
\description{
Implements logistic regression with a continuous response.
}
\details{
- INCLUDE note on deviance (not comparable between lin and log models)
- alpha: elastic net parameter\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)
- do not use "family"
}
\examples{
n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)

### Add ... to all glmnet::glmnet calls !!! ###

}
