% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cornet}
\alias{cornet}
\alias{cornet-package}
\title{Combined regression}
\usage{
cornet(y, cutoff, X, alpha = 1, npi = 101, pi = NULL, nsigma = 99,
  sigma = NULL, nfolds = 10, foldid = NULL,
  type.measure = "deviance", ...)
}
\arguments{
\item{y}{continuous outcome\strong{:}
vector of length \eqn{n}}

\item{cutoff}{cutoff point for dichotomising outcome into classes\strong{:}
\emph{meaningful} value between \code{min(y)} and \code{max(y)}}

\item{X}{features\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{alpha}{elastic net mixing parameter\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{npi}{number of \code{pi} values (weighting)}

\item{pi}{pi sequence\strong{:}
vector of values in the unit interval;
or \code{NULL} (default sequence)}

\item{nsigma}{number of \code{sigma} values (scaling)}

\item{sigma}{sigma sequence\strong{:}
vector of increasing positive values;
or \code{NULL} (default sequence)}

\item{nfolds}{number of folds}

\item{foldid}{fold identifiers\strong{:}
vector with entries between \eqn{1} and \code{nfolds};
or \code{NULL} (balance)}

\item{type.measure}{loss function for binary classification
(linear regression uses the deviance)}

\item{...}{further arguments passed to \code{\link[glmnet]{glmnet}}}
}
\value{
Returns an object of class \code{cornet}, a list with multiple slots:
\itemize{
   \item \code{gaussian}: fitted linear model, class \code{glmnet}
   \item \code{binomial}: fitted logistic model, class \code{glmnet}
   \item \code{sigma}: scaling parameters \code{sigma},
          vector of length \code{nsigma}
   \item \code{pi}: weighting parameters \code{pi},
          vector of length \code{npi}
   \item \code{cvm}: evaluation loss,
          matrix with \code{nsigma} rows and \code{npi} columns
   \item \code{sigma.min}: optimal scaling parameter,
          positive scalar
   \item \code{pi.min}: optimal weighting parameter,
          scalar in unit interval
   \item \code{cutoff}: threshold for dichotomisation
}
}
\description{
Implements logistic regression for dichotomised outcomes 
in high-dimensional settings.
}
\details{
This function fits a \emph{gaussian} model for the numeric response,
and a \emph{binomial} model for the binary response,
meaning that the \code{glmnet} argument \code{family} is unavailable.
Even if \code{type.measure} equals \code{"deviance"},
the loss is incomparable between linear and logistic regression.
}
\examples{
n <- 100; p <- 200
y <- rnorm(n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
net <- cornet(y=y,cutoff=0,X=X)
net

}
\references{
A Rauschenberger, E Glaab (2019).
"Lasso and ridge regression with dichotomised outcomes".
\emph{Manuscript in preparation}.
}
\seealso{
Methods for objects of class \code{cornet} include
\code{\link[=coef.cornet]{coef}} and
\code{\link[=predict.cornet]{predict}}.
}
