% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{drop.trivial}
\alias{drop.trivial}
\title{Drop trivial test}
\usage{
drop.trivial(map)
}
\arguments{
\item{map}{list with names "genes", "exons", and "snps"
(output from \code{map.genes}, \code{map.exons}, and \code{map.snps})}
}
\description{
This function
}
\details{
This functions drops tests for genes without SNPs or with a single exon.
}
\examples{
NA

}
