% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.exons.bbmri}
\alias{get.exons.bbmri}
\title{Get exon data (BBMRI)}
\usage{
get.exons.bbmri(path = getwd())
}
\arguments{
\item{path}{data directory}
}
\description{
This function transforms exon data (virtual machine):
loads quality controlled gene expression data,
extracts sample identifiers,
removes samples without SNP data,
loads exon expression data,
extracts sample identifiers,
retains samples that passed quality control,
retains exons on the autosomes
}
\examples{
path <- "/virdir/Scratch/arauschenberger/trial"

}
