% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.snps.bbmri}
\alias{get.snps.bbmri}
\title{Get SNP data (BBMRI)}
\usage{
get.snps.bbmri(chr, biobank = NULL, path = getwd(), size = 500 * 10^3)
}
\arguments{
\item{chr}{chromosome: integer \eqn{1-22}}

\item{biobank}{character "CODAM", "LL", "LLS", "NTR", "PAN", "RS", or NULL (all)}

\item{path}{data directory}

\item{size}{maximum number of SNPs to read in at once;
trade-off between memory usage (low) and speed (high)}
}
\description{
This function transforms SNP data (virtual machine):
limits analysis to specified biobanks,
reads in genotype data in chunks,
removes SNPs with missing values (multiple biobanks/technologies),
removes SNPs with low minor allele frequency,
fuses data from multiple biobanks/technologies
}
\examples{
path <- "/virdir/Scratch/arauschenberger/trial"

}
