% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.snps.bbmri}
\alias{get.snps.bbmri}
\title{Get SNP data (BBMRI)}
\usage{
get.snps.bbmri(chr, biobank = NULL, path = getwd())
}
\arguments{
\item{chr}{chromosome: integer \eqn{1-22}}

\item{biobank}{character "CODAM", "LL", "LLS", "NTR", "PAN", "RS", or NULL (all)}

\item{path}{data directory}
}
\description{
This function transforms SNP data (virtual machine).
}
\examples{
path <- "/virdir/Scratch/arauschenberger/trial"

}
