% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grid}
\alias{grid}
\title{Plot grid}
\usage{
grid(x, y, xlim = NULL, ylim = NULL, n = 10)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{xlim}{horizontal limits}

\item{ylim}{vertical limits}

\item{n}{grid resolution}
}
\description{
This function tests for alternative splicing.
}
\examples{
x <- stats::rbeta(n=100,shape1=0.2,shape2=0.5)
y <- stats::rbeta(n=100,shape1=0.2,shape2=0.5)
grid(x,y)

}
