% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{joinet}
\alias{joinet}
\title{Multivariate Elastic Net Regression}
\usage{
joinet(
  Y,
  X,
  family = "gaussian",
  nfolds = 10,
  foldid = NULL,
  type.measure = "deviance",
  alpha.base = 1,
  alpha.meta = 1,
  weight = NULL,
  sign = NULL,
  ...
)
}
\arguments{
\item{Y}{outputs\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{q} columns (outputs)}

\item{X}{inputs\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (inputs)}

\item{family}{distribution\strong{:}
vector of length \eqn{1} or \eqn{q} with entries
\code{"gaussian"}, \code{"binomial"} or \code{"poisson"}}

\item{nfolds}{number of folds}

\item{foldid}{fold identifiers\strong{:}
vector of length \eqn{n} with entries between \eqn{1} and \code{nfolds};
or \code{NULL} (balance)}

\item{type.measure}{loss function\strong{:}
vector of length \eqn{1} or \eqn{q} with entries
\code{"deviance"}, \code{"class"}, \code{"mse"} or \code{"mae"}
(see \code{\link[glmnet]{cv.glmnet}})}

\item{alpha.base}{elastic net mixing parameter for base learners\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{alpha.meta}{elastic net mixing parameter for meta learners\strong{:}
numeric between \eqn{0} (ridge) and \eqn{1} (lasso)}

\item{weight}{input-output effects\strong{:}
matrix with \eqn{p} rows (inputs) and \eqn{q} columns (outputs)
with entries \eqn{0} (exclude) and \eqn{1} (include),
or \code{NULL} (see details)}

\item{sign}{output-output effects\strong{:}
matrix with \eqn{q} rows ("meta-inputs") and \eqn{q} columns (outputs), 
with entries \eqn{-1} (negative), \eqn{0} (none),
\eqn{1} (positive) and \eqn{NA} (any),
or \code{NULL} (see details)}

\item{...}{further arguments passed to \code{\link[glmnet]{glmnet}}}
}
\value{
This function returns an object of class \code{joinet}.
Available methods include
\code{\link[=predict.joinet]{predict}},
\code{\link[=coef.joinet]{coef}},
and \code{\link[=weights.joinet]{weights}}.
The slots \code{base} and \code{meta} each contain
\eqn{q} \code{\link[glmnet]{cv.glmnet}}-like objects.
}
\description{
Implements multivariate elastic net regression.
}
\details{
TO BE DELETED:
constraint
non-negativity constraints\strong{:}
logical (see details)
non-negativity constraints\strong{:}
If it is reasonable to assume that the outcomes
are \emph{positively} correlated
(potentially after changing the sign of some outcomes)
we recommend to set \code{constraint=TRUE}.
Then non-negativity constraints are imposed on the meta learner.

\strong{input-output relations:}
In this matrix with \eqn{p} rows and \eqn{q} columns,
the entry in the \eqn{j}th row and the \eqn{k}th column
indicates whether the \eqn{j}th input may be used for 
modelling the \eqn{k}th output
(where \eqn{0} means yes and
\eqn{1} means no).
By default, 

\strong{output-output relations:}
In this matrix with \eqn{q} rows and \eqn{q} columns,
the entry in the \eqn{k}th row and the \eqn{l}th column
indicates how the \eqn{k}th output may be used for
modelling the \eqn{l}th output
(where \eqn{-1} means negative effect,
\eqn{0} means no effect,
\eqn{1} means positive effect,
and \eqn{NA} means any effect).
All entries on the diagonal must equal \eqn{1}.
By default (\code{sign=NULL}),
Kendall correlation determines this matrix,
with \eqn{-1} for significant negative, \eqn{0} for insignificant,
\eqn{1} for significant positive correlations.
The short-cuts \code{sign=1} and \code{code=NA} set all off-diagonal entries
equal to \eqn{1} (non-negativity constraints)
or \code{NA} (no constraints), respectively.
The former short-cut is useful if all pairs of outcomes
are assumed to be \emph{positively} correlated
(potentially after changing the sign of some outcomes).

\strong{elastic net:}
\code{alpha.base} controls input-output effects,
\code{alpha.meta} controls output-output effects;
lasso renders sparse models (\code{alpha}\eqn{=1}),
ridge renders dense models (\code{alpha}\eqn{=0})
}
\examples{
\dontshow{
if(!grepl('SunOS',Sys.info()['sysname'])){
n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
object <- joinet(Y=Y,X=X)}}
\dontrun{
n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
object <- joinet(Y=Y,X=X)}

\dontrun{
browseVignettes("joinet") # further examples}

}
\references{
Armin Rauschenberger, Enrico Glaab (2021)
"Predicting correlated outcomes from molecular data"
\emph{Manuscript in preparation}.
}
\seealso{
\code{\link{cv.joinet}}, vignette
}
