% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{map.exons}
\alias{map.exons}
\title{Search for exons}
\usage{
map.exons(gene_id, exon_id)
}
\arguments{
\item{gene_id}{gene names\strong{:} vector with one entry per gene (gene names!)}

\item{exon_id}{exon names\strong{:} vector with one entry per exon
(also gene names! separated by comma if multiple genes)}
}
\description{
This function
}
\details{
The exon names should contain the gene names. For each gene, this function
returns the indices of the exons.
}
\examples{
NA

}
