% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{map.snps}
\alias{map.snps}
\title{Search for SNPs}
\usage{
map.snps(gene.chr, gene.start, gene.end, snp.chr, snp.pos, dist = 10^3)
}
\arguments{
\item{gene.chr}{chromosome\strong{:}
numeric vector with one entry per gene}

\item{gene.start}{start position\strong{:}
numeric vector with one entry per gene}

\item{gene.end}{end position\strong{:}
numeric vector with one entry per gene}

\item{snp.chr}{integer 1-22}

\item{snp.pos}{chromosomal position of SNPs\strong{:}
numeric vector with one entry per SNP}

\item{dist}{number of base pairs before start position\strong{:}
integer}
}
\description{
This function
}
\examples{
gene.chr <- rep(1,times=5)
gene.start <- 1:5
gene.end <- 2:6

snp.chr <- rep(1,times=100)
snp.pos <- seq(from=1,to=4.9,length.out=100)

map.snps(gene.chr,gene.start,gene.end,snp.chr,snp.pos,dist=0)

}
