% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.cornet}
\alias{plot.cornet}
\title{Plot loss matrix}
\usage{
\method{plot}{cornet}(x, ...)
}
\arguments{
\item{x}{\link[cornet]{cornet} object}

\item{...}{further arguments (not applicable)}
}
\description{
Plots the loss for difference combinations of
the weight (pi) and scale (sigma) paramters.
}
\examples{
NA

}
