% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test.multiple}
\alias{test.multiple}
\title{Conduct multiple tests}
\usage{
test.multiple(Y, X, map, rho = c(0, 0.5, 1))
}
\arguments{
\item{Y}{exon expression\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (exons)}

\item{X}{SNP genotype\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{q} columns (SNPs)}

\item{map}{list with names "genes", "exons", and "snps"
(output from \code{map.genes}, \code{map.exons}, and \code{map.snps})}

\item{rho}{correlation\strong{:}
numeric vector with values between \eqn{0} and \eqn{1}}
}
\description{
This function ...
}
\details{
Automatic adjustment of the number of permutations
such that Bonferroni-significant p-values are possible.
}
\examples{
NA

}
