% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{visualise}
\alias{visualise}
\title{Plot SNP-exon correlations}
\usage{
visualise(Y, X, map, i)
}
\arguments{
\item{Y}{to do}

\item{X}{to do}

\item{map}{to do}

\item{i}{to do}
}
\description{
This function ...
}
\examples{
# see vignette

}
